@@ -24,16 +24,17 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "@angular/animations": "^8.2.14",
+    "@angular/animations": "^9.0.2",
     "@angular/cdk": "^8.2.3",
-    "@angular/common": "^8.2.14",
-    "@angular/compiler": "^8.2.14",
-    "@angular/core": "^8.2.14",
-    "@angular/forms": "^8.2.14",
+    "@angular/common": "^9.0.2",
+    "@angular/compiler": "^9.0.2",
+    "@angular/core": "^9.0.2",
+    "@angular/forms": "^9.0.2",
     "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "^8.2.14",
-    "@angular/platform-browser-dynamic": "^8.2.14",
-    "@angular/router": "^8.2.14",
+    "@angular/platform-browser": "^9.0.2",
+    "@angular/platform-browser-dynamic": "^9.0.2",
+    "@angular/router": "^9.0.2",
+    "@fullcalendar/core": "^4.4.0",
     "@mdi/font": "^3.8.95",
     "@ngrx/effects": "^8.6.0",
     "@ngrx/router-store": "^8.6.0",
@@ -53,24 +54,24 @@
     "node-localstorage": "^1.3.1",
     "path": "^0.12.7",
     "primeflex": "^1.0.0",
-    "primeicons": "^1.0.0",
-    "primeng": "^8.1.1",
+    "primeicons": "^2.0.0",
+    "primeng": "^9.0.0-rc.4",
     "quill": "^1.3.7",
     "rxjs": "6.5.4",
     "sessionstorage": "^0.1.0",
     "tslib": "^1.10.0",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.25",
-    "@angular/cli": "^8.3.25",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/language-service": "^8.2.14",
-    "@angular/platform-server": "^8.2.14",
+    "@angular-devkit/build-angular": "~0.900.3",
+    "@angular/cli": "^9.0.3",
+    "@angular/compiler-cli": "^9.0.2",
+    "@angular/language-service": "^9.0.2",
+    "@angular/platform-server": "^9.0.2",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "^2.0.8",
-    "@types/node": "~8.9.4",
-    "codelyzer": "^5.2.1",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^4.4.1",
@@ -83,7 +84,7 @@
     "ts-loader": "^5.4.5",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~3.5.3",
+    "typescript": "~3.7.5",
     "webpack-cli": "^3.3.10"
   },
   "prettier": {
