@@ -21,14 +21,14 @@ import { SplitGutterDirective } from './components/split-gutter.directive';
 })
 export class AngularSplitModule {
 
-    public static forRoot(): ModuleWithProviders {
+    public static forRoot(): ModuleWithProviders<AngularSplitModule> {
         return {
             ngModule: AngularSplitModule,
             providers: []
         };
     }
 
-    public static forChild(): ModuleWithProviders {
+    public static forChild(): ModuleWithProviders<AngularSplitModule> {
         return {
             ngModule: AngularSplitModule,
             providers: []
