@@ -54,7 +54,7 @@ export class AppSkinCoreModule {
    * Получить модуль с провайдерами для корня приложения.
    * @returns {ModuleWithProviders} Модуль с провайдерами.
    */
-  static forRoot(): ModuleWithProviders {
+  static forRoot(): ModuleWithProviders<AppSkinCoreModule> {
     return {
       ngModule: AppSkinCoreModule,
       providers: [
