@@ -40,7 +40,7 @@ export class AppSkinHostModule {
    * Получить модуль с провайдерами для корня приложения.
    * @returns {ModuleWithProviders} Модуль с провайдерами.
    */
-  static forRoot(): ModuleWithProviders {
+  static forRoot(): ModuleWithProviders<AppSkinHostModule> {
     return {
       ngModule: AppSkinHostModule,
       providers: []
