@@ -43,7 +43,7 @@ export class AppSkinRootModule {
    * Получить модуль с провайдерами для корня приложения.
    * @returns {ModuleWithProviders} Модуль с провайдерами.
    */
-  static forRoot(): ModuleWithProviders {
+  static forRoot(): ModuleWithProviders<AppSkinRootModule> {
     return {
       ngModule: AppSkinRootModule,
       providers: []
