@@ -45,7 +45,7 @@ export class AppCoreModule {
    * Получить модуль с провайдерами для корня приложения.
    * @returns {ModuleWithProviders} Модуль с провайдерами.
    */
-  static forRoot(): ModuleWithProviders {
+  static forRoot(): ModuleWithProviders<AppCoreModule> {
     return {
       ngModule: AppCoreModule,
       providers: [
