@@ -34,7 +34,7 @@ export class AppDataModule {
    * Получить модуль с провайдерами для корня приложения.
    * @returns {ModuleWithProviders} Модуль с провайдерами.
    */
-  static forRoot(): ModuleWithProviders {
+  static forRoot(): ModuleWithProviders<AppDataModule> {
     return {
       ngModule: AppDataModule,
       providers: []
