@@ -59,7 +59,7 @@ export class AppRootModule {
    * Получить модуль с провайдерами для корня приложения.
    * @returns {ModuleWithProviders} Модуль с провайдерами.
    */
-  static forRoot(): ModuleWithProviders {
+  static forRoot(): ModuleWithProviders<AppRootModule> {
     return {
       ngModule: AppRootModule,
       providers: [
